removed embeds on %give and %take because they were unnecessary
i replaced the messages with ones similar to the previous set which were
Just Fine. Upstream clearly had no qualms since they merged it in so
they can keep it if they wish. And they'll probably complain to me that
given the chance, I didn't complain. But half the point of me doing this
separate branch is because i don't want to wait around on the pr process
anymore to write code

whomst would use "I" in UI messages anyway? Maybe it works in Harold's
mind, but it just looks awkward to just about everyone else.......

i should change the remote that `development` pushes to...